move constraint class
