+++ sql/catalyst/src/main/java/org/apache/spark/sql/connector/catalog/constraints/Constraint.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.sql.connector.catalog;
+package org.apache.spark.sql.connector.catalog.constraints;
 
 import org.apache.spark.sql.connector.expressions.filter.Predicate;
