+++ sql/core/src/test/scala/org/apache/spark/sql/execution/command/v2/CheckConstraintSuite.scala
@@ -18,8 +18,8 @@
 package org.apache.spark.sql.execution.command.v2
 
 import org.apache.spark.sql.{AnalysisException, QueryTest}
-import org.apache.spark.sql.connector.catalog.Constraint.Check
 import org.apache.spark.sql.connector.catalog.Table
+import org.apache.spark.sql.connector.catalog.constraints.Constraint.Check
 import org.apache.spark.sql.execution.command.DDLCommandTestUtils
 
 class CheckConstraintSuite extends QueryTest with CommandSuiteBase with DDLCommandTestUtils {
